Restructure sqs folders
