+++ backend-api/src/functions/adapters/sqs/types.ts
@@ -1,0 +1,6 @@
+export type IdCheckSqsMessages = VendorProccessingMessage;
+
+interface VendorProccessingMessage {
+  biometricSessionId: string;
+  sessionId: string;
+}
